Create images_1054
Request for add the Tajweed Color Quran (NON SPACED VERSION) to your app? The official PDF (high resolution) is available from the below page:

https://ar.islamway.net/book/15763/%D9%85%D8%B5%D8%AD%D9%81-%D8%A7%D9%84%D8%AA%D8%AC%D9%88%D9%8A%D8%AF-%D8%A7%D9%84%D9%85%D9%84%D9%88%D9%86

This is a excellent quran, and the 4th edition (which does not have the missing kasrah Page 457 / Chapter 38 / Ayah 71 in the word Lilmalaika).

This would be my dream wish if Quran Android can support this Mushaf, as it is Othman font. Can you add this feature, or maybe allow custom pdf import support?